# backend/docs/occacia-openapi-3.1.yml
openapi: 3.1.0
info:
  title: Occacia API
  version: "1.0.0"
  description: >
    Occacia API contract aligned with use cases, EERD, and class diagram.
    Core flow: Event -> Tasks -> Recommendations -> Packages -> Confirm Package -> Package Order -> Vendor Fulfillment Requests -> Task execution.
servers:
  - url: https://app.occacia.com/api/v1
    description: Production (example)
  - url: http://localhost:3000/api/v1
    description: Local dev (example)
security:
  - bearerAuth: []
...
paths:
...
  /auth/admin/register:
    post:
      tags: [Authentication]
      summary: Provision admin account (internal)
      description: Restricted/internal endpoint; not self-service.
      responses:
        "201":
          description: Admin provisioned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegisterResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /auth/admin/login:
    post:
      tags: [Authentication]
      summary: Admin login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in (tokens issued)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /admin/vendors:
    get:
      tags: [Admin]
      summary: List vendors (management)
      parameters:
        - name: approval_status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/VendorApprovalStatus" }
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/AccountStatus" }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedVendors" }
  /admin/vendors/{vendorId}:
    parameters:
      - $ref: "#/components/parameters/VendorId"
    get:
      tags: [Admin]
      summary: Get vendor detail (management)
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VendorAdminView" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/vendors/{vendorId}/status:
    parameters:
      - $ref: "#/components/parameters/VendorId"
    put:
      tags: [Admin]
      summary: Update vendor account status (ACTIVE/SUSPENDED/DISABLED)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VendorStatusUpdateRequest" }
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VendorAdminView" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /admin/vendors/{vendorId}/approve:
    parameters:
      - $ref: "#/components/parameters/VendorId"
    post:
      tags: [Admin]
      summary: Approve vendor registration
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VendorAdminView" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /admin/vendors/{vendorId}/reject:
    parameters:
      - $ref: "#/components/parameters/VendorId"
    post:
      tags: [Admin]
      summary: Reject vendor registration
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VendorRejectRequest" }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VendorAdminView" }
        "400": { $ref: "#/components/responses/BadRequest" }

  # =========================
  # Admin - Customer Management (UC-29)
  # =========================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
...
    VendorId:
      name: vendorId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
...
  responses:
    BadRequest:
      description: Bad request / validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    UUID:
      type: string
      format: uuid
    ISODateTime:
      type: string
      format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
...
    UserRole:
      type: string
      enum: [CUSTOMER, VENDOR, ADMIN]
    AccountStatus:
      type: string
      enum: [ACTIVE, PENDING, SUSPENDED, DISABLED]
    VendorApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    OrganizationStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
...
    RegisterResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user:
          type: object
          required: [userId, role, status, email]
          properties:
            userId: { $ref: "#/components/schemas/UUID" }
            email: { type: string, format: email }
            role: { $ref: "#/components/schemas/UserRole" }
            status: { $ref: "#/components/schemas/AccountStatus" }

    # Profiles
...
    Organization:
      type: object
      required: [orgId, name, code, status, createdAt, updatedAt]
      properties:
        orgId: { $ref: "#/components/schemas/UUID" }
        name: { type: string }
        code: { type: string, description: "Unique join code" }
        status: { $ref: "#/components/schemas/OrganizationStatus" }
        kymDetails: { type: object, additionalProperties: true }
        createdAt: { $ref: "#/components/schemas/ISODateTime" }
        updatedAt: { $ref: "#/components/schemas/ISODateTime" }
...
    VendorAdminView:
      type: object
      required: [vendorId, email, displayName, approvalStatus, status]
      properties:
        vendorId: { $ref: "#/components/schemas/UUID" }
        email: { type: string, format: email }
        displayName: { type: string }
        contactPhone: { type: string }
        approvalStatus: { $ref: "#/components/schemas/VendorApprovalStatus" }
        approvedAt: { $ref: "#/components/schemas/ISODateTime" }
        organizationId: { $ref: "#/components/schemas/UUID" }
        organization: { $ref: "#/components/schemas/Organization" }
        status: { $ref: "#/components/schemas/AccountStatus" }
    PaginatedVendors:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/VendorAdminView" }
        nextCursor: { type: string }
    VendorStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/AccountStatus" }
    VendorRejectRequest:
      type: object
      properties:
        reason: { type: string }
...